+++ realm/realm-library/src/objectServer/java/io/realm/SyncUser.java
@@ -74,6 +74,11 @@ public class SyncUser {
                                         user.getIdentity(),
                                         error.toString()));
                             }
+
+                            @Override
+                            public void onClientResetRequired(SyncSession session, ClientResetHandler handler) {
+                                RealmLog.error("Client Reset required for users management Realm: " + user.toString());
+                            }
                         })
                         .modules(new PermissionModule())
                         .build();
@@ -293,13 +298,6 @@ public class SyncUser {
                 }
             }
 
-            // Stop all active sessions immediately. If we waited until after talking to the server
-            // there is a high chance errors would be reported from the Sync Client first which would
-            // be confusing.
-            for (SyncSession session : sessions) {
-                session.getOsSession().stop();
-            }
-
             SyncManager.getUserStore().remove(syncUser.getIdentity());
 
             // Delete all Realms if needed.
@@ -390,9 +388,9 @@ public class SyncUser {
      *
      * @return the user's access token. If this user has logged out or the login has expired {@code null} is returned.
      */
-    public String getAccessToken() {
+    public Token getAccessToken() {
         Token userToken = syncUser.getUserToken();
-        return (userToken != null) ? userToken.value() : null;
+        return (userToken != null) ? userToken : null;
     }
 
     /**